+++ crengine/src/lvtinydom.cpp
@@ -454,7 +454,7 @@ lUInt32 calcGlobalSettingsHash(int documentId, bool already_rendered)
     hash = hash * 31 + (int)fontMan->GetHintingMode();
     hash = hash * 31 + LVRendGetBaseFontWeight();
     hash = hash * 31 + fontMan->GetFallbackFontFaces().getHash();
-    hash = hash * 31 + gRenderDPI;
+    hash = hash * 31 + LVRendGetRenderDPI();
     // If not yet rendered (initial loading with XML parsing), we can
     // ignore some global flags that have not yet produced any effect,
     // so they can possibly be updated between loading and rendering